requirements.txt
Django==4.2.20
Pillow==11.2.1
django-crispy-forms==2.4
crispy-bootstrap5==2025.4
pillow-heif==0.22.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9 